readable HostPaths, generalize Readable caching
* host paths can now be read, provided they don't attempt to escape the
  context dir
* Readable now implements CachePath which can be used to return a
  host-side file path containing the Readable's contents
